Update to build and test on Windows, including x86_64
The change in date.d allows compilation on both Win64 and Win32.

The change in stream.d is not obvious but tempCStringW() corrupts
the filename where tempCString!wchar() works. This is possibly a
generic bug in std.string

Simple change in doformat.d to call the local copy of doFormat
and get rid of the deprecation warning in Phobos

Added the win64.make alongside win32, although dub allows building
and testing Win64 already.

Plus some extra .gitignore stuff for VisualD and dub.